# ARROW-11033 [Rust] Csv writing performance improvements 