--- a/rust/arrow/src/csv/writer.rs
+++ b/rust/arrow/src/csv/writer.rs
@@ -77,14 +77,11 @@ const DEFAULT_DATE_FORMAT: &str = "%F";
 const DEFAULT_TIME_FORMAT: &str = "%T";
 const DEFAULT_TIMESTAMP_FORMAT: &str = "%FT%H:%M:%S.%9f";
 
-unsafe fn vector_as_slice<'a, T>(buf: &'a mut Vec<T>) -> &'a mut [T] {
-    let first = buf.as_mut_ptr();
-    std::slice::from_raw_parts_mut(first, buf.capacity())
-}
 pub fn to_string<N: lexical_core::ToLexical>(n: N) -> String {
     let mut buf = Vec::<u8>::with_capacity(N::FORMATTED_SIZE_DECIMAL);
     unsafe {
-        let len = lexical_core::write(n, vector_as_slice(&mut buf)).len();
+        let mut slice = std::slice::from_raw_parts_mut(first.as_mut_ptr(), buf.capacity());
+        let len = lexical_core::write(n, slice).len();
         buf.set_len(len);
         String::from_utf8_unchecked(buf)
     }