--- a/rust/arrow/src/csv/writer.rs
+++ b/rust/arrow/src/csv/writer.rs
@@ -80,7 +80,7 @@ const DEFAULT_TIMESTAMP_FORMAT: &str = "%FT%H:%M:%S.%9f";
 pub fn to_string<N: lexical_core::ToLexical>(n: N) -> String {
     let mut buf = Vec::<u8>::with_capacity(N::FORMATTED_SIZE_DECIMAL);
     unsafe {
-        let mut slice = std::slice::from_raw_parts_mut(first.as_mut_ptr(), buf.capacity());
+        let mut slice = std::slice::from_raw_parts_mut(buf.as_mut_ptr(), buf.capacity());
         let len = lexical_core::write(n, slice).len();
         buf.set_len(len);
         String::from_utf8_unchecked(buf)