--- a/rust/arrow/src/csv/writer.rs
+++ b/rust/arrow/src/csv/writer.rs
@@ -86,7 +86,7 @@ pub fn to_string<N: lexical_core::ToLexical>(n: N) -> String {
         //  Soundness
         //      Length of buf is set as written length afterwards. lexical_core
         //      creates a valid string, so doesn't need to be checked.
-        let mut slice = std::slice::from_raw_parts_mut(buf.as_mut_ptr(), buf.capacity());
+        let slice = std::slice::from_raw_parts_mut(buf.as_mut_ptr(), buf.capacity());
         let len = lexical_core::write(n, slice).len();
         buf.set_len(len);
         String::from_utf8_unchecked(buf)