Some performance improvements for the csv writer

* Use lexical core for numeric types
* Allow setting batch size in convert (slightly faster reading when using higher batch size, could also be helpful in writing)
* Avoid allocation of vec


PR: `cargo run --release --bin tpch -- convert --input path --output ./output --format csv -s 20000`
Orders / lineitems:

```
Conversion completed in 2050 ms
Conversion completed in 16955 ms
```
Master `cargo run --release --bin tpch -- convert --input path --output ./output --format csv`

```
Conversion completed in 2336 ms
Conversion completed in 19070 ms
```